refactor: co-locate task form variable utilities
