+++ webapp/client/apps/orchestration-cluster-webapp/src/tasklist/modules/task-details-form/formatVariablesToFormData.ts
